\name{GO.groups}
\docType{data}
\alias{GO.groups}
\title{GO terms and genes}
\usage{
 data(yeast)
}
\description{
  This dataset contains all Gene Ontology (GO) terms and their associated genes, derived from S. cerevisiae annotations. 
}
\format{
 A list with 3837 entries.
}
\source{
The Gene Ontology Consortium. Gene ontology: tool for the unification of biology. Nat. Genet.. May 2000;25(1):25-9.
}

