\name{mean_invgamma}
\alias{mean_invgamma}
\title{The mean of the inverse gamma distribution}
\usage{
mean_invgamma(alpha, beta)
}
\arguments{
  \item{alpha}{the shape parameter of the inverse gamma
  distribution}

  \item{beta}{the scale parameter of the inverse gamma
  distribution}
}
\value{
the mean value of the parameterized inverse gamma
disribution, given by beta/(alpha-1) if alpha > 1
}
\description{
This function returns the mean of inverse gamma distributed
random variables
}

