\name{rnorminvgamma}
\alias{rnorminvgamma}
\title{Draw normal-inverse-gamma distributed random variates}
\usage{
rnorminvgamma(n, alpha, beta)
}
\arguments{
  \item{n}{the number of variates to generate}

  \item{alpha}{shape parameter of the inverse gamma
  distribution}

  \item{beta}{scale parameter of the inverse gamma
  distribution}
}
\value{
A vector of random variates arising form the
normal-inverse-gamma distribution
}
\description{
This function will return normal-inverse-gamma distributed
random variates
}

