\name{yeast.homozygote}
\docType{data}
\alias{yeast.homozygote}
\title{Yeast homozygote transcription profiles}
\usage{
 data(yeast)
}
\description{
  This dataset has normalized log-fold-change in RNA levels between 3 yeast species and S. cerevisiae. Generated by performing RNAseq on homozygotes of each species to quantify expression.
}
\format{
 A matrix with 4 rows and 4835 columns. Each row is a species and each column is a gene.
}
\source{
 Schraiber, et al (2013). Inferring evolutionary histories of pathway regulation from transcriptional profiling data. PLoS Computational Biology 9:e10003255.
}

