\name{yeast.hybrid}
\docType{data}
\alias{yeast.hybrid}
\title{Yeast hybrid transcription profiles}
\usage{
 data(yeast)
}
\description{
  This dataset has normalized log-fold-change in RNA levels between 3 yeast species and S. cerevisiae, generated by creating hybrids of each species with S. cerevisiae and using RNAseq to quantify allele specific expression.
}
\format{
 A matrix with 4 rows and 4835 columns. Each row is a species and each column is a gene.
}
\source{
 Schraiber, et al (2013). Inferring evolutionary histories of pathway regulation from transcriptional profiling data. PLoS Computational Biology 9:e10003255.
}

