\name{yeast.tree}
\docType{data}
\alias{yeast.tree}
\title{Phylogenetic tree of yeast species}
\usage{
 data(yeast)
}
\description{
  Ape-format ultrametric phylogenetic tree of the yeast species used in Schraiber, et al (2013).
}
\format{
 ape tree object
}
\source{
 Schraiber, et al (2013). Inferring evolutionary histories of pathway regulation from transcriptional profiling data. PLoS Computational Biology 9:e10003255.
}

